feat: Embedded provider core
