+++ packages/browser-sdk/package.json
@@ -31,6 +31,7 @@
     "@phantom/base64url": "workspace:^",
     "@phantom/browser-injected-sdk": "workspace:^",
     "@phantom/client": "workspace:^",
+    "@phantom/embedded-provider-core": "workspace:^",
     "@phantom/parsers": "workspace:^",
     "axios": "^1.10.0",
     "bs58": "^6.0.0",
