+++ packages/browser-sdk/src/providers/embedded/adapters/url-params.ts
@@ -1,0 +1,16 @@
+import { URLParamsAccessor } from "@phantom/embedded-provider-core";
+
+/**
+ * Browser implementation using window.location.search
+ */
+export class BrowserURLParamsAccessor implements URLParamsAccessor {
+  getParam(key: string): string | null {
+    const urlParams = new URLSearchParams(window.location.search);
+    return urlParams.get(key);
+  }
+}
+
+/**
+ * Default accessor instance for browser environments
+ */
+export const browserUrlParamsAccessor: URLParamsAccessor = new BrowserURLParamsAccessor();
