+++ packages/embedded-provider-core/.eslintignore
@@ -1,0 +1,8 @@
+node_modules
+.eslintrc.js
+dist/
+**/dist/
+**/node_modules/
+/**/node_modules/*
+build/
+**/build/
