+++ packages/embedded-provider-core/jest.config.js
@@ -1,0 +1,6 @@
+const sharedJestConfig = require("../../sharedJestConfig");
+
+module.exports = {
+  ...sharedJestConfig,
+  displayName: "@phantom/embedded-provider-core",
+};
