+++ packages/embedded-provider-core/src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from "./storage";
+export * from "./url-params";
+export * from "./auth";
+export * from "./platform";
