+++ packages/embedded-provider-core/src/interfaces/url-params.ts
@@ -1,11 +1,11 @@
 /**
  * URL parameter accessor abstraction
- * 
+ *
  * This allows us to abstract URL parameter access for different environments:
  * - Browser: window.location.search via URLSearchParams
  * - React Native: deep links or webview callback URLs
  * - Mobile WebView: in-app browser navigation state
- * 
+ *
  * This abstraction makes the auth flow code portable across platforms
  * without needing to mock window.location directly in tests.
  */
@@ -17,18 +17,3 @@ export interface URLParamsAccessor {
    */
   getParam(key: string): string | null;
 }
-
-/**
- * Browser implementation using window.location.search
- */
-export class BrowserURLParamsAccessor implements URLParamsAccessor {
-  getParam(key: string): string | null {
-    const urlParams = new URLSearchParams(window.location.search);
-    return urlParams.get(key);
-  }
-}
-
-/**
- * Default accessor instance for browser environments
- */
-export const urlParamsAccessor: URLParamsAccessor = new BrowserURLParamsAccessor();
