+++ packages/embedded-provider-core/src/utils/session.ts
@@ -1,0 +1,9 @@
+export function generateSessionId(): string {
+  return (
+    "session_" +
+    Math.random().toString(36).substring(2, 15) +
+    Math.random().toString(36).substring(2, 15) +
+    "_" +
+    Date.now()
+  );
+}
